+++ pp/components/bend_circular_heater.py
@@ -1,10 +1,10 @@
 import pp
 from pp.cell import cell
 from pp.component import Component
+from pp.config import TECH
 from pp.cross_section import cross_section
 from pp.path import arc, extrude
 from pp.snap import snap_to_grid
-from pp.tech import TECH
 
 
 @cell